+++ scripts/install/enginescript-install.sh
@@ -378,6 +378,19 @@ fi
 print_last_errors
 debug_pause "NTP"
 
+# DigitalOcean Remote Console (optional)
+if [[ "${INSTALL_DIGITALOCEAN_REMOTE_CONSOLE}" = "1" ]]; then
+  if [[ "${DO_CONSOLE}" = 1 ]];
+    then
+      echo "DigitalOcean Remote Console script has already run."
+    else
+      /usr/local/bin/enginescript/scripts/install/system-misc/digitalocean-console-install.sh 2>> /tmp/enginescript_install_errors.log
+      echo "DO_CONSOLE=1" >> /var/log/EngineScript/install-log.log
+  fi
+  print_last_errors
+  debug_pause "DigitalOcean Remote Console"
+fi
+
 # PCRE
 if [[ "${PCRE}" = 1 ]];
   then